Introduce more items and some bug fix

+++ Assets/Scripts/UI/ExperienceController.cs
@@ -8,17 +8,17 @@ using TMPro;
 public class ExperienceController : MonoBehaviour {
     public int _level;
     private float _expMultiplier;
-    [SerializeField] private float _currentMaxExp;
-    [SerializeField] private float _currentExp;
+    [SerializeField] private int _currentMaxExp;
+    [SerializeField] private int _currentExp;
     [SerializeField] private TMP_Text _expText;
     [SerializeField] private TMP_Text _levelText;
     [SerializeField] private Slider slider;
     public UnityEvent OnLevelUp;
     
-    public void SetPlayerInitialMaxExp(float maxExp) {
+    public void SetPlayerInitialMaxExp(int maxExp) {
         _level = 1;
         slider.maxValue = maxExp;
-        _currentExp = 0f;
+        _currentExp = 0;
         _currentMaxExp = maxExp;
         _expMultiplier = 0.05f * maxExp; 
         slider.value = 0f;
@@ -26,7 +26,7 @@ public class ExperienceController : MonoBehaviour {
         UpdateLevelText();
     }
 
-    public void IncreaseExp(float amountExp) {
+    public void IncreaseExp(int amountExp) {
         _currentExp += amountExp;
         slider.value += amountExp;
         UpdateExpText();
@@ -43,7 +43,7 @@ public class ExperienceController : MonoBehaviour {
     public void LevelUp() {
         _currentExp = _currentExp % _currentMaxExp;
         slider.maxValue = Mathf.Floor(_currentMaxExp * _expMultiplier + _level);
-        _currentMaxExp = slider.maxValue;
+        _currentMaxExp = (int)slider.maxValue;
         slider.value = _currentExp;
         _level ++;
         UpdateExpText();
